pyproject.toml
[tool.poetry]
name = "police-for-university-study"
version = "0.1.0"
description = ""
authors = [""]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
python-decouple = "^3.8"
psycopg2 = "^2.9.10"
dj-database-url = "^2.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
